+++ flutter/lib/src/default_integrations.dart
@@ -367,8 +367,15 @@ class LoadReleaseIntegration extends Integration<SentryFlutterOptions> {
       if (!kIsWeb) {
         if (options.release == null || options.dist == null) {
           final packageInfo = await _packageLoader();
+          var name = packageInfo.packageName;
+          if (name.isEmpty) {
+            // Not all platforms have a packageName.
+            // If no packageName is available, use the appName instead.
+            name = _cleanAppName(packageInfo.appName);
+          }
+
           final release =
-              '${packageInfo.packageName}@${packageInfo.version}+${packageInfo.buildNumber}';
+              '$name@${packageInfo.version}+${packageInfo.buildNumber}';
           options.logger(SentryLevel.debug, 'release: $release');
 
           options.release = options.release ?? release;
@@ -381,5 +388,17 @@ class LoadReleaseIntegration extends Integration<SentryFlutterOptions> {
     }
 
     options.sdk.addIntegration('loadReleaseIntegration');
+  }
+
+  String _cleanAppName(String appName) {
+    // Replace disallowed chars with an underscore '_'
+    // https://docs.sentry.io/platforms/flutter/configuration/releases/#bind-the-version
+    return appName
+        .replaceAll('/', '_')
+        .replaceAll('\\', '_')
+        .replaceAll('\t', '_')
+        .replaceAll('\r\n', '_')
+        .replaceAll('\r', '_')
+        .replaceAll('\n', '_');
   }
 }